Pull review fixes
