+++ App.tsx
@@ -4,14 +4,20 @@ import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import { Ionicons } from '@expo/vector-icons';
-import FavPokemon from './components/FavPokemon';
-import PokemonList from './components/PokemonList';
-import NativeModule from './components/NativeModule';
+import FavPokemon from './screens/FavPokemon';
+import PokemonList from './screens/PokemonList';
+import NativeModule from './screens/NativeModule';
 
 const Stack = createStackNavigator();
 const TabNav = createBottomTabNavigator();
 
 export default function App() {
+  enum TabScreenNames {
+    FAV_POKEMON = 'FavPokemon',
+    POKEMON_LIST = 'PokemonList',
+    NATIVE_MODULE = 'NativeModule'
+  }
+
   return (
     <NavigationContainer>
       <TabNav.Navigator
@@ -19,7 +25,7 @@ export default function App() {
           tabBarIcon: ({ focused, color, size }) => {
             let iconName;
 
-            if (route.name === 'FavPokemon') {
+            if (route.name === TabScreenNames.FAV_POKEMON) {
               iconName = 'md-heart';
             } else if (route.name === 'PokemonList') {
               iconName = 'md-paw';
@@ -34,44 +40,32 @@ export default function App() {
           inactiveTintColor: 'gray'
         }}
       >
-        <TabNav.Screen name='FavPokemon' component={FavPokemon} />
-        <TabNav.Screen name='PokemonList' component={PokemonList} />
-        <TabNav.Screen name='NativeModule' component={NativeModule} />
+        <TabNav.Screen
+          name={TabScreenNames.FAV_POKEMON}
+          component={FavPokemon}
+        />
+        <TabNav.Screen
+          name={TabScreenNames.POKEMON_LIST}
+          component={PokemonList}
+        />
+        <TabNav.Screen
+          name={TabScreenNames.NATIVE_MODULE}
+          component={NativeModule}
+        />
       </TabNav.Navigator>
       {/* <Stack.Navigator initialRouteName='FavPokemon'>
         <Stack.Screen
-          name='FavPokemon'
+          name={TabScreenNames.FAV_POKEMON}
           component={FavPokemon}
           options={{ title: 'My Favourite Pokemon' }}
         />
         <Stack.Screen
-          name='PokemonList'
+          name={TabScreenNames.POKEMON_LIST}
           component={PokemonList}
           options={{ title: 'List Of Pokemons' }}
         />
-        <Stack.Screen name='NativeModule' component={NativeModule} />
+        <Stack.Screen name={TabScreenNames.NATIVE_MODULE} component={NativeModule} />
       </Stack.Navigator> */}
     </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: 'aqua',
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  image: {
-    backgroundColor: 'yellow',
-    width: 150,
-    height: 150,
-    marginTop: 30,
-    marginBottom: 10
-  },
-  header: {
-    fontSize: 25,
-    fontWeight: '700',
-    color: 'orange'
-  }
-});
